Payroll: Allow passing scope to PublicCredentials.
When using the Payroll API, you are required to supply
the 'scope' argument to select which parts of the API you
will require access to.

This is not 100% simple, as it requires them in a comma-
separated list, but urlencoding this was problematic.

The 'FULL_API' feature is not yet complete.