Add Payroll API support.
We have several new datetime and date fields, as well as some
boolean and multi-line fields. Added a comment about what a
multi-line field is, for clarity.

Since the Payroll API uses a different endpoint, we need to be
able to change this. Passing it in when constructing the Manager
seems a reasonable solution.