Add Payroll API support.
Add the URL endpoint, and the list of object types.

Note there is a change to the 'Organisation' object list,
as the API is inconsistent here, and this was the only
way to actually get at the data.

This is incomplete: I'm still in the process of developing
against this API, so haven't really tested the access of objects
other than Employees, PayrollCalendars.