Change the process for converting data from XML to python.
Okay, this is the big one.

Firstly, I skip fields that are empty XML tags.

I've added a new field type (DECIMAL_FIELDS), although it looks like
I've commented that bit out :)

The big difference though is that I do stuff in a bit of a different
order.

Firstly, I look at ones that are a 'value' type thing, and convert those
to the primitive they should be. Otherwise, we recursively call the
method to convert all children.

Then, and this is the biggest difference, I think, we ignore the number
of elements, and only check to see if it is a collection.

This check occurs later (in the other branch of the if statement),
so that _anything_ that is in MULTI_LINES is always a list in the end.