Change the process for generating XML.
Because with some MULTI_LINES objects we might actually end up
with a list of primitives, we need to handle that slightly differently.

If we don't have a dict in our dict_to_xml, we simply set the text of
the root element to the value we got passed in.

When we are pushing data back to the server, we turn the data into a
list, and then run the process. I'm not convinced about this, but it
is working so far :(