Request data in JSON format.
This replaces a lot of the parsing code: it no longer needs to
deserialize from XML, which removes some of the complexity of the
conversion. (Resolves #8).

It also means that collection-type objects always return a collection,
even if there is only one object (Resolves #9).

This has implications for 'get', in that this should only return the
one object (as it has been accessed via it's id).