Prevent attempt to re-read rsa_key file.
When a given process has already read the file
into a string, it does not need to re-read it:
indeed, it will fail (as the key data is in the
variable, rather than the filename).

This is only a quick fix: it should make a decision
about reading the file on startup, or whenever it
is required (probably the latter).